GH-1176: extract context/ into internal sub-package
Move all YAML schema types, project context assembly, prompt document
structures, and constitution rendering from pkg/orchestrator/ into
pkg/orchestrator/internal/context/. Parent files become thin wrappers
with type aliases and function delegates preserving the existing API.
